Re-organize directories to conform to updated schema
