--- source/modules/data_processing/math.py
@@ -1,4 +1,0 @@
-# ============================================================================
-# == Functions == #
-def number_in_range(x, minimum, maximum):
-    return minimum <= x <= maximum
