--- source/modules/data_processing/prompts.py
@@ -1,10 +1,0 @@
-def yes_no(prompt):
-    while True:
-        user_input = input(f"{prompt} (Y/N) ")
-        if user_input == "Y":
-            return True
-        elif user_input == "N":
-            return False
-        else:
-            print("Invalid input...")
-            continue
